restructure and cleanup
shared/components => components/common
