+++ src/Client/Components/Common/CustomValidation.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
 
-namespace FSH.BlazorWebAssembly.Client.Shared;
+namespace FSH.BlazorWebAssembly.Client.Components.Common;
 
 // See https://docs.microsoft.com/en-us/aspnet/core/blazor/forms-validation?view=aspnetcore-6.0#server-validation-with-a-validator-component
 public class CustomValidation : ComponentBase
