+++ src/Client/Components/Common/TablePager.razor
