+++ src/Client/Components/EntityTable/AddEditModal.razor.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel.DataAnnotations;
+using FSH.BlazorWebAssembly.Client.Components.Common;
 using FSH.BlazorWebAssembly.Client.Shared;
 using Microsoft.AspNetCore.Components;
 using MudBlazor;
