+++ src/Client/Pages/Authentication/ForgotPassword.razor.cs
@@ -1,4 +1,5 @@
-﻿using FSH.BlazorWebAssembly.Client.Infrastructure.ApiClient;
+﻿using FSH.BlazorWebAssembly.Client.Components.Common;
+using FSH.BlazorWebAssembly.Client.Infrastructure.ApiClient;
 using FSH.BlazorWebAssembly.Client.Shared;
 using FSH.WebApi.Shared.Multitenancy;
 using Microsoft.AspNetCore.Components;
