+++ src/Client/Shared/ApiHelper.cs
@@ -1,4 +1,5 @@
-﻿using FSH.BlazorWebAssembly.Client.Infrastructure.ApiClient;
+﻿using FSH.BlazorWebAssembly.Client.Components.Common;
+using FSH.BlazorWebAssembly.Client.Infrastructure.ApiClient;
 using MudBlazor;
 
 namespace FSH.BlazorWebAssembly.Client.Shared;
