--- src/Client/Shared/Error.razor
@@ -1,23 +1,0 @@
-﻿@using Microsoft.Extensions.Logging
-@inject ILogger<Error> Logger
-
-<CascadingValue Value=this>
-    @ChildContent
-</CascadingValue>
-
-@code {
-    [Parameter]
-    public RenderFragment ChildContent { get; set; } = default!;
-
-    public void ProcessError(Exception ex)
-    {
-        Snackbar.Add(ex.Message, Severity.Error);
-    }
-    public void ProcessError(List<string> messages)
-    {
-        foreach (var message in messages)
-        {
-            Snackbar.Add(message, Severity.Error);
-        }
-    }
-}
