+++ src/Client/_Imports.razor
@@ -16,9 +16,8 @@
 @using FSH.BlazorWebAssembly.Client.Shared
 @using FSH.BlazorWebAssembly.Client.Pages.Authentication
 @using FSH.BlazorWebAssembly.Client.Pages.Multitenancy
-@using FSH.BlazorWebAssembly.Client.Shared.Components
+@using FSH.BlazorWebAssembly.Client.Components.Common
 @using FSH.BlazorWebAssembly.Client.Components.EntityTable
-@using FSH.BlazorWebAssembly.Client.Components.ErrorHandler
 @using FSH.BlazorWebAssembly.Client.Components.Localization
 @using FSH.BlazorWebAssembly.Client.Components.Notifications
 @using FSH.BlazorWebAssembly.Client.Components.ThemeManager
